Check for hidden _method on form for remote validation
When creating a form in Laravel, the form may have a hidden input with a
name of '_method' to describe the actual method that should be used. The
existing validation was checking the 'METHOD' attribute of the form
element, which is not necessarily correct. Modified code to check for
hidden _method attribute and use that instead if it exists.