Fix array wildcard validation in Laravel 5.4 (#297)
I know it is a very dirty fix, but my OOP knowledge doesn't allow me to do it in a better way. At least, I have found a problem. 
The problem is addRules() method in new Laravel Validator  class. It uses ValidationRuleParser to transform all the rules with * into explicitAttributes and explicitRules properties, that will be set based on real data ( * => 0, 1, 2 etc), but we don't have any real data when we try to grab set of rules from Request class. And these explicitAttributes are not added to 'rules' array that is used in JsValidatior to build rules. 
So, the main idea to change * with something else to trick ValidationRuleParser and then change it back.